[BugFix]Fix Conversion warning for gcc(#12071) (#12072)
Problem:
If we enable Wconversion for compiling in gcc, we will get a lot warning for it. This issue is mainly remove part of warning from the code.

Solution:
We will use more reasonable variable type in the code, except for:
1. The modification will cause performance drop.
2. Some small-length variables come from third-party libraries, so we can only ensure that our code is safe when we used it.
3. Some functions logically need to do this type conversion.
4. From the semantics of the function, there will be no overflow.

+++ be/src/util/debug_util.h
@@ -50,6 +50,6 @@ std::string get_short_version();
 // Returns "<program short name> version <GetBuildVersion(compact)>"
 std::string get_version_string(bool compact);
 
-std::string hexdump(const char* buf, int len);
+std::string hexdump(const char* buf, size_t len);
 
 } // namespace starrocks
